+++ src/js/sidenav.js
@@ -12,8 +12,12 @@ class Sidenav {
   constructor(element, options) {
     this.defaultOptions = {
       overlay: true,
-      bodyScrolling: false
+      bodyScrolling: false,
+      animation: 'slide-in',
+      animationDelay: 300
     };
+
+    this.animationList = ['slide-in', 'push'];
 
     this.el = document.querySelector(element);
     this.el.Sidenav = this;
@@ -21,6 +25,7 @@ class Sidenav {
     this.isActive = false;
     this.isFixed = this.el.classList.contains('fixed');
     this.isLarge = this.el.classList.contains('large');
+    this.isDefaultAnimation = true;
 
     /**
      * Options
@@ -29,12 +34,9 @@ class Sidenav {
      * @property {boolean} bodyScrolling Prevent bodyScrolling when sidenav is active and over content
      */
     this.options = extend(this.defaultOptions, options);
+    this.options.animation = this.options.animation.toLowerCase();
 
-    if (this.options.overlay) {
-      this._createOverlay();
-    }
     this._setup();
-    this.el.classList.contains('large') ? document.body.classList.add('sidenav-large') : '';
   }
 
   /**
@@ -47,8 +49,12 @@ class Sidenav {
       }
     });
     if (this.options.overlay) {
+      this._createOverlay();
       this.overlayElement.addEventListener('click', e => this._onClickTrigger(e, this.el.id));
     }
+    this.el.classList.contains('large') ? document.body.classList.add('sidenav-large') : '';
+    this.animationList.includes(this.options.animation) ? this._handleAnim() : '';
+    this.el.style.transitionDuration = this.options.animationDelay + 'ms';
   }
 
   /**
@@ -61,16 +67,33 @@ class Sidenav {
   }
 
   /**
+   * Add .anim-{name} to body
+   */
+  _handleAnim() {
+    if (this.options.animation !== 'slide-in') {
+      document.body.classList.add('anim-' + this.options.animation);
+      document.body.style.transitionDuration = this.options.animationDelay + 'ms';
+      this.isDefaultAnimation = false;
+    }
+  }
+
+  /**
    * Enable or disable body scroll when option is true
    * @param {boolean} state Enable or disable body scroll
    */
   _toggleBodyScroll(state) {
     if (!this.options.bodyScrolling) {
-      if (state) {
-        document.body.style.overflow = '';
-      } else {
-        document.body.style.overflow = 'hidden';
-      }
+      state ? (document.body.style.overflow = '') : (document.body.style.overflow = 'hidden');
+    }
+  }
+
+  /**
+   * Add .anim-active to body
+   * @param {boolean} state Enable or disable animation
+   */
+  _toggleAnim(state) {
+    if (!this.isDefaultAnimation) {
+      state ? document.body.classList.add('anim-active') : document.body.classList.remove('anim-active');
     }
   }
 
@@ -99,6 +122,7 @@ class Sidenav {
     this.el.classList.add('active');
     this.overlay(true);
     this._toggleBodyScroll(false);
+    this._toggleAnim(true);
   }
 
   /**
@@ -107,7 +131,10 @@ class Sidenav {
   close() {
     this.el.classList.remove('active');
     this.overlay(false);
-    this._toggleBodyScroll(true);
+    this._toggleAnim(false);
+    setTimeout(() => {
+      this._toggleBodyScroll(true);
+    }, this.options.animationDelay);
   }
 
   /**
@@ -116,11 +143,7 @@ class Sidenav {
    */
   overlay(state) {
     if (this.options.overlay) {
-      if (state) {
-        document.body.appendChild(this.overlayElement);
-      } else {
-        document.body.removeChild(this.overlayElement);
-      }
+      state ? document.body.appendChild(this.overlayElement) : document.body.removeChild(this.overlayElement);
     }
   }
 }